+++ node/src/components/block_synchronizer/config.rs
@@ -10,7 +10,6 @@ const DEFAULT_PEER_REFRESH_INTERVAL: &str = "90sec";
 const DEFAULT_NEED_NEXT_INTERVAL: &str = "1sec";
 const DEFAULT_DISCONNECT_DISHONEST_PEERS_INTERVAL: &str = "10sec";
 const DEFAULT_LATCH_RESET_INTERVAL: &str = "5sec";
-const DEFAULT_STALL_LIMIT: &str = "120sec";
 
 /// Configuration options for fetching.
 #[derive(Copy, Clone, DataSize, Debug, Deserialize, Serialize)]
@@ -25,9 +24,6 @@ pub struct Config {
     pub disconnect_dishonest_peers_interval: TimeDiff,
     /// Time interval for resetting the latch in block builders.
     pub latch_reset_interval: TimeDiff,
-    /// Time interval after which synchronization is considered stalled if no successful sync
-    /// activity happened.
-    pub stall_limit: TimeDiff,
 }
 
 impl Default for Config {
@@ -41,7 +37,6 @@ impl Default for Config {
             )
             .unwrap(),
             latch_reset_interval: TimeDiff::from_str(DEFAULT_LATCH_RESET_INTERVAL).unwrap(),
-            stall_limit: TimeDiff::from_str(DEFAULT_STALL_LIMIT).unwrap(),
         }
     }
 }
